refectoring-2

+++ panimation.cpp
@@ -3,7 +3,7 @@
 
 void PAnimation::draw(PEventLoop * loop,std::shared_ptr<SDL_Renderer> renderer){
     if(_path){
-        if(_path_index < _path->points().size()){
+        if(static_cast<std::size_t>(_path_index) < _path->points().size()){
             bool point_arrived = go_to(_path->points().at(_path_index), loop);
             if(point_arrived){
                 animate(_path->points().at(_path_index), loop);
